char: msm_smd_pkt: Reduce log severity of signals
If the user process is interrupted by a signal while the smd_pkt driver
is executing a wait on behalf of the client, this activity gets logged as
an error.  Since this scenario may be expected, and generally can be
transparent to the client as VFS may restart the syscall, the log level
of this event should be reduced.

Reduce log events related to signal handling as INFO so that they can be
filtered out as appropriate.

Change-Id: I2ee59556a96ebc44bd499511b84b0c80b7922179
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>